services/conversion/requirements.txt
pika~=1.3.2
pillow~=11.1.0
wand~=0.6.13
python-magic~=0.4.27
rawpy~=0.24.0
numpy~=2.2.1
imageio
pydub
audioop-lts
pypandoc
python-docx
reportlab
fpdf
PyPDF2
pdf2image
pytesseract~=0.3.13
ebooklib
pdf2docx
beautifulsoup4~=4.12.3
PyMuPDF
htmldocx~=0.0.6
html2docx~=1.6.0
pdf2docx
pikepdf
pdfplumber~=0.11.5
pdfservices-sdk==4.1.0
fonttools~=4.55.3
setuptools
brotli
patool~=3.1.0
pyunpack~=0.3
pyexcel
pyexcel-xls
pyexcel-xlsx
pandas
xlwt
openpyxl
tablib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests